Add a FIFO IP and compressing logic for writeback.
The FIFO IP was pre-generated by Vivado and is configured to have a
128-bit width, a depth of 128, and with lookahead enabled. A wrapper is
provided to merge multiple FIFOs to a single, large FIFO.

The compressing logic assumes data streams in as a tuple {din,
din_valid, din_width}. din is the actually data input, din_valid shows
whether din is valid, and din_width indicates that the first din_width
bits of din is valid. The compressing logic compress the valid data and
streams them out.