Go back to in-memory vivado project to parrallel synthesis
I was wrong about how to inspecting designs in the vivado GUI. Leaving
project related files on filesystem is not a straightforward way.
Besides, it even makes parallel synthesis harder.

So I want to revert a previous change and use in-memory project again.
To inspect synthesized designs, I can open the generated (intermediate)
checkpoints in the vivado GUI.