Add axi_register_slice(_light) to the cl_fpgarr_wrapper
This is one of the many trials I do in the hope of getting rid of timing
violations.
The timing violations I see is related to SH_DDR, etc. instead of what I
add (cl_fpgarr_*). One guess is that lack of register_slice in
cl_fpgarr_wrapper could force the pnr prioritize the wrapper to be
placed closer to the shell, which breaks the timing of the underlying
application.