Add a skeleton of the AXI-based backend trace storage module
Rename rr_writeback_bus_t to rr_stream_bus_t. This unifies the data bus
used in both record and replay.

Add an axil interconnect to split the bar1 (2MB) to two parts.
The lower 1MB is pass through to the cl.
The higher 1MB is reserved for rr_cfg_bus.

The content of the storage backend module is actually a placeholder that
does nothing.