Overhaul the packing module to support customized merging strategy
Due to the timing violation in pnr, I have a hypothesis that the merging
tree structure is suboptimal. Based on the observation of the
schematics, customizable merging stragies are now supported in the
packing module via constant parameter array.

The replay part also uses the SHUFFLE_PLAN.
i.e. the trace decoder will use the shuffled CHANNEL_WIDTHS to determine
the actual width of each logging unit.

See the implementation notes for more details.