Trace dump and bug fix.
Add software-side support to dump the collected trace to memory. Now
during make we can pass an environment variable RR_MODE={record,replay}.
When in record mode, the simulator dumps the trace to a file named
record.dump; when in replay mode, the simulator loads the trace from the
aforementioned file and programs the hardware to read it.