Refactor GET_LEN used in the storage backend
GET_LEN is a constant function used in the storage backend to determine
the length of a logging unit, which is then used to parse the
record/replay trace.
Due to the function scope limitation (within a module, no inherence) and
the error-prone way of duplicating function definition,
this commit:

1. Add a MACRO to define the GET_LEN function and pass the
shuffled channel widths parameter array down from the wrapper.

2. Change GET_LEN signature to only take LOGB_CHANNEL_CNT width data as
input.

3. Lower the SHUFFLED_CHANNEL_WIDTHS definition from the wrapper to
tracestorage to workaround vivado synthesis errors.
That error means I should not use hierarchical expression to define
localparam. i.e.
> cannot resolve hierarchical name for the item