Try to improve timing: Add pipe stages to csr
This decision was made according to the timing report, where the
rt_replay_pkt_cnt was placed in a different SLR from the rest of the
rr_parse_replay_trace module.
The reason is suspected to be the fact that the counter is passed to the
rr_csrs directly via wire.