Add almful_hi (for PCIM-W only) and almful_lo (for all other channels)
The almful_hi is implemented as the programmable full signal in xpm_fifo_sync.
The almful_lo is implemented as the programmable empty signal in xpm_fifo_sync.

merged_fifo can no longer serve as the fifo used in the happenbefore_encoder.

Again, this is to prevent the deadlock among pcim interconnect, cl_pcim
and logging writeback.