Split rr_state to oneoff states and real-time states, expose more almful
oneoff states are those which should be accumulated by logic OR, like
underflow and overflow. I only care if they ever happened or not.

real-time states are those which should be directly passed to the CSR,
like almful. I care the real-time value, like whether almful is asserted
when the circuit gets stuck.