Update the new buggy axis_fifo example to use both ocl and pcis
I want to trigger an axis_fifo overflow on real hardware. The current
plan is to make two threads, one thread pours in 32-bit ocl request and
the other thread do 512-bit pcis write requests. Both of them would be
merged and bufferred into the buggy axis_fifo before being written to
the on-FPGA DRAM. The hypothesis is that whenever the axis_fifo is full
and the size of all ocl and pcis requests does not align with the buffer
size (e.g. 1x32-bit ocl + 2x512-bit pcis does not align with a
1024x32-bit buffer), there will be buffer overflow and packet loss.
Later the cpu-side would read back those requests and check whether
there was a packet loss.

Cannot trigger the bug in simulation because:
1. the framework would segfault if use multi-thread
2. the axis_fifo drains too fast in simulation, cannot force
Need to test this on real hardware.