[fpgarr_sv] Register another APP internal axi bus for RR
This is to demonstrate the flexibility and reusability of Vidi.
In this commit, I exemplify how to pick an application internal AXI-like
interface and prepare it for RR.
Note that the changes also enable this new interface to be configured to
temporary include/exclude in RR via the autogen py scripts.

Have tested on VCS simulation that the logging_bus of this internal AXI
bus has meaningful data.